Move collecting logic to server and update gci
